Fix warnings, replace some static inline with inline + extern inline, and fix glibc support

+++ memrcasechr.c
@@ -7,7 +7,7 @@ void *
 libsimple_memrcasechr(const void *s_, int c, size_t n_)
 {
 	char *s = *(char **)(void *)&s_;
-	ssize_t n = n_;
+	ssize_t n = (ssize_t)n_;
 	c = tolower(c);
 	while (n-- && tolower(s[n]) != c);
 	return n < 0 ? NULL : &s[n];
